Added a transition from old way of saving bars.
Previously, only the name of each bar present was saved. But now, we are saving
an id with each, so we can identify each bar uniquely even after reordering.

This commit adds an angular run function that changes the bars' saved format so
the rest of the extension works smoothly.